requirements.txt
typer>=0.9.0
ffmpeg-python>=0.2.0
pytube>=15.0.0
yt-dlp>=2023.12.30
anthropic>=0.17.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
Pillow>=10.0.0